Add files via upload
Making more prompts if the input is incorrect as well as adding a re-run option which has greatly helped testing. Next step (you can see current progress at the bottom) is to make a few more methods so that the Main() is a lot more tidy. After that I think it is pretty solid.